refactor: split commands into one per cog, started on state system
- commands are now split one per file which will help maintain a cleanly and modularized codebase
- began working on a state system so that the bot can be ran in multiple guilds at once simultaneously
- cleaned up large chunks of the codebase
- somethings are still broken atm, don't expect too much :D
