+++ djkhaled/state.py
@@ -1,0 +1,4 @@
+from collections import defaultdict
+
+song_queue = {}
+voice_clients = {}
